% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDesc.R
\docType{data}
\name{bcos}
\alias{bcos}
\title{Breast Cosmesis Data}
\format{
An object of class \code{data.frame} with 94 rows and 3 columns.
}
\usage{
bcos
}
\value{
Breast cosmesis data contains the following fields:
\item{left}{a numeric vector}
\item{right}{a numeric vector}
\item{treatment}{a factor with levels Rad and RadChem}
}
\description{
The often used data set for interval censored data, described and given in full in Finkelstein and Wolfe (1985).
}
\examples{
data = data(bcos)
}
\references{
Finkelstein D.M. and Wolfe R.A.(1985). "A semiparametric model for regression analysis of interval-censored failure time data." \emph{Biometrics} \strong{41}, 933-945.
}
\keyword{datasets}
