% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDesc.R
\docType{data}
\name{cadi}
\alias{cadi}
\title{The children’s absenteeism data in Indonesia}
\format{
An object of class \code{data.frame} with 437 rows and 2 columns.
}
\usage{
cadi
}
\value{
The children’s absenteeism data in Indonesia contains the following fields:
\item{y1}{The number of days the children missed their primary activities due to illness in the last four weeks}
\item{y2}{The number of days the children spent in bed due to illness in the last four weeks}
}
\description{
In a survey of Indonesian family life conducted by Strauss et al. the participants included 7000 households sampled from 321 communities randomly selected from 13 of the nation’s 26 Provinces,
in which 83\% of the Indonesian population lived. Among those households with one child per household, 437 household heads were asked questions about the health of their children.
}
\examples{
data = data(cadi)
}
\references{
Huang X.F., Tian G.L., Zhang, C. and Jiang, X.(2017). "Type I multivariate zero-inflated generalized Poisson distribution with applications." \emph{Statistics and its Interface} \strong{10}(2), 291-311.

Strauss J., Beegle K., Sikoki B., Dawiyanto A., Herawati Y. and Witoelar Y.(2004). "The Third Wave of the Indonesia Family Life Survey (IFLS): Overview and Field Report, WR-144/1-NIA/NICHD, RAND Corporation, Santa Monica, CA."
}
\keyword{datasets}
