% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoxProfile.R
\name{CoxProfile}
\alias{CoxProfile}
\title{Title}
\usage{
CoxProfile(x, I, t, be, la, n, q)
}
\arguments{
\item{x}{a data frame containing the variables of the model}

\item{I}{The status indicator, normally 0=alive, 1=dead.}

\item{t}{for right censored data}

\item{be}{a vector of regression parameters}

\item{la}{baseline hazard rate}
}
\value{
requires a value
}
\description{
Title
}
\examples{
x = matrix(rnorm(10*1,-1,1),10,1)
la = 2; be = 0.5
u=runif(10); t=-log(u)/(la*exp(x\%*\%be))
cen=3.7; I=1*(t<=cen)
CoxProfile(x, I, t, be, la)
}
