% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COXsample.R
\name{COXsample}
\alias{COXsample}
\title{Generate sample from COX model}
\usage{
COXsample(n, be, la)
}
\arguments{
\item{n}{Number of samples generated}

\item{be}{Coefficients for COX regression}

\item{la}{Baseline lambda for COX regression (we take constant value here)}
}
\value{
List with length n including covariate x, censoring indicator I and event time t
}
\description{
Generate sample from COX model
}
