% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcp.R
\name{curvature}
\alias{curvature}
\title{Inferred curvature of an \code{mmad_expr}}
\usage{
curvature(expr)
}
\arguments{
\item{expr}{An \code{mmad_expr}.}
}
\value{
A length-1 character.
}
\description{
Returns the DCP curvature of the expression as one of \code{"affine"},
\code{"convex"}, \code{"concave"}, or \code{"unknown"}. \code{"unknown"} means the DCP
rule set could not prove a definite curvature; this is a refusal,
not a claim of non-convexity.
}
\examples{
curvature(log(mmad_var(1) + 1))                      # "concave"
curvature(exp(mmad_var(1)))                          # "convex"
curvature(2 * mmad_var(1) + 3 * mmad_var(2))         # "affine"
curvature(exp(mmad_var(1)) - exp(mmad_var(2)))       # "unknown"
}
