% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CZIGP_CRate.R
\name{CZIGP_CRate}
\alias{CZIGP_CRate}
\title{Title}
\usage{
CZIGP_CRate(phi0, phi, la, th)
}
\arguments{
\item{phi0}{probability}

\item{phi}{probability}

\item{la}{scale parameter}

\item{th}{discrete parameter}
}
\value{
requires a value
}
\description{
Title
}
\examples{
CZIGP_CRate(0.5,0.5,1,0.1)
}
