% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CZIGPsample.R
\name{CZIGP_sample}
\alias{CZIGP_sample}
\title{Generate sample from Multivariate Zero-inflated Compound Generalized Poisson Distribution}
\usage{
CZIGP_sample(phi0, phi, la, th, n)
}
\arguments{
\item{phi0}{Probability of taking value 0 for outer Bernoulli distribution}

\item{phi}{Vector of probability of taking value 0 for inner Bernoulli distributions}

\item{la}{Lambda for Poison distribution}

\item{th}{Theta for Poison distribution}

\item{n}{Number of samples generated}
}
\value{
Vector of samples with dimension n
}
\description{
Generate sample from Multivariate Zero-inflated Compound Generalized Poisson Distribution
}
