% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CZIGPFx.R
\name{CZIGPFx}
\alias{CZIGPFx}
\title{Title}
\usage{
CZIGPFx(phi0, phi, la, th)
}
\arguments{
\item{phi0}{probability}

\item{phi}{probability}

\item{la}{scale parameter}

\item{th}{discrete parameter}
}
\value{
requires a value
}
\description{
Title
}
\examples{
CZIGPFx(0.5,0.5,1,0.1)
}
