% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CZIGPmm.R
\name{CZIGPmm}
\alias{CZIGPmm}
\title{MM algorithm based on the AD method for Multivariate compound zero-inflated generalized Poisson distribution}
\usage{
CZIGPmm(
  y,
  phi0,
  phi,
  la,
  th,
  Maxiter = 500,
  convergence = 1e-06,
  method = "ADMM",
  ...
)
}
\arguments{
\item{y}{matrix data}

\item{phi0}{probability}

\item{phi}{probability}

\item{la}{scale parameter}

\item{th}{discrete parameter}
}
\value{
requires a value
}
\description{
MM algorithm based on the AD method for Multivariate compound zero-inflated generalized Poisson distribution
}
\examples{
y = matrix(7,1,1);
CZIGPmm(0.5,0.5,1,0.1,y)
}
