% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CZIGPQx.R
\name{CZIGPQx}
\alias{CZIGPQx}
\title{Title}
\usage{
CZIGPQx(phi0, phi, la, th)
}
\arguments{
\item{phi0}{probability}

\item{phi}{probability}

\item{la}{scale parameter}

\item{th}{discrete parameter}
}
\value{
requires a value
}
\description{
Title
}
\examples{
CZIGPQx(0.5,0.5,1,0.1)
}
