% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GaFrailtyMM.R
\name{GaFrailtyMM}
\alias{GaFrailtyMM}
\title{The MM algorithm based on AD technology for gamma frailty model}
\usage{
GaFrailtyMM(
  formula,
  data,
  beta = NULL,
  theta = NULL,
  lambda = NULL,
  Maxiter = 1000,
  convergence = 1e-06,
  ...
)
}
\arguments{
\item{formula}{A formula object, which contains on the left hand side an object of the type \code{Surv}
and on the right hand side a \code{+cluster(id)} statement.
e.g. \code{formula=Surv(time, status) ~ x + cluster(id)}}

\item{data}{A \code{data.frame} in which to interpret the variables named in the formula.}

\item{beta}{A vector of unknown regression parameters, default is \code{NULL}.
If is \code{NULL}, then make all \code{beta=1} during calculation.}

\item{theta}{The Variance of frailty factors subject to Gamma distribution, default is \code{NULL}.
If is \code{NULL}, then let \code{theta=1} during calculation.}

\item{lambda}{Baseline hazard rate, default set to \code{NULL}. If is \code{NULL},
then let each \code{lambda} equals to \code{1/N} during calculation, which \code{N} is the number of observed.}

\item{Maxiter}{The maximum number of iterations; default=1000.}

\item{convergence}{Specify the convergence criterion, the default is 1e-6.}

\item{...}{Additional arguments}
}
\value{
An object of class \code{GaFrailtyMM} that contains the following fields:
\item{N}{list of total observations}
\item{TotalCen}{list of censored observations}
\item{print_k}{The iterations}
\item{print_err}{Convergence result}
\item{ELL}{The Log Likelihood value}
\item{th}{\eqn{\theta}}
\item{std_th}{The standard deviation of the estimated \eqn{\theta}}
\item{ci_th_lower, ci_th_upper}{The likelihood-based 95\% confidence interval for the \eqn{\theta}}
\item{be}{\eqn{\beta}}
\item{std_be}{The standard deviation of the estimated \eqn{\beta}}
\item{ci_be_lower, ci_be_upper}{The likelihood-based 95\% confidence interval for the \eqn{\beta}}
\item{convergence}{The final convergence value}
\item{namesX}{The variable name}
}
\description{
The MM algorithm based on AD technology for gamma frailty model
}
\examples{
gf = GaFrailtyMM(Surv(time, status) ~ age + sex + cluster(id), kidney)
summary.GaFrailtyMM(gf)
}
