% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IC2mm.R
\name{IC2mm}
\alias{IC2mm}
\title{MM algorithm based on the AD method for Case II interval-censored data}
\usage{
IC2mm(L, R, Maxiter = 500, convergence = 1e-06, method = "ADMM", ...)
}
\arguments{
\item{L}{Left interval boundary}

\item{R}{Right interval boundary}
}
\value{

}
\description{
MM algorithm based on the AD method for Case II interval-censored data
}
\examples{
L = c(0.5, 2); R = c(1, 3)
IC2mm(L, R)
}
