% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IC2sample.R
\name{IC2sample}
\alias{IC2sample}
\title{Generate sample for Type II interval-censored data}
\usage{
IC2sample(la, n)
}
\arguments{
\item{la}{Hazard rate for observed data}

\item{n}{Number of samples generated}
}
\value{
List of n objects containing left and right bounds for observed data
}
\description{
Generate sample for Type II interval-censored data
}
