% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LTN_CRate.R
\name{LTN_CRate}
\alias{LTN_CRate}
\title{Title}
\usage{
LTN_CRate(a, mu, sig2)
}
\arguments{
\item{a}{left border}

\item{mu}{mean}

\item{sig2}{variance}
}
\value{

}
\description{
Title
}
\examples{
a = 5; mu = 0.5; si2 = 0.5
LTN_CRate(a, mu, si2)
}
