% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LTNsample.R
\name{LTN_sample}
\alias{LTN_sample}
\title{Generate sample from Left Truncated Normal Distribution}
\usage{
LTN_sample(a, mu, si2, n)
}
\arguments{
\item{a}{Left truncation point}

\item{mu}{Distribution mean for normal distribution before left truncation}

\item{si2}{Distribution variance for normal distribution before left truncation}

\item{n}{Number of samples generated}
}
\value{
Vector of samples with dimension n
}
\description{
Generate sample from Left Truncated Normal Distribution
}
