% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LTNFx.R
\name{LTNFx}
\alias{LTNFx}
\title{Title}
\usage{
LTNFx(a, mu, si2)
}
\arguments{
\item{a}{left border}

\item{mu}{mean}

\item{si2}{variance}
}
\value{

}
\description{
Title
}
\examples{
a = 5; mu = 0.5; si2 = 0.5
LTNFx(a,mu,si2)
}
