% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LTNmm.R
\name{LTNmm}
\alias{LTNmm}
\title{MM algorithm based on the AD method for Left-truncated normal distribution}
\usage{
LTNmm(a, mu, si2, y)
}
\arguments{
\item{a}{left border}

\item{mu}{mean}

\item{si2}{variance}

\item{y}{data}
}
\value{
data
}
\description{
MM algorithm based on the AD method for Left-truncated normal distribution
}
\examples{
LTNmm(5, 7, 4, 3)
y=c(3, 9); LTNmm(5, 7, 4, y)
}
