% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LTNQx.R
\name{LTNQx}
\alias{LTNQx}
\title{Title}
\usage{
LTNQx(a, mu, si2)
}
\arguments{
\item{a}{left border}

\item{mu}{mean}

\item{si2}{variance}
}
\value{
requires a value
}
\description{
Title
}
\examples{
a = 5; mu = 0.5; si2 = 0.5
LTNQx(a, mu, si2)
}
