% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr.R
\name{Math.mmad_expr}
\alias{Math.mmad_expr}
\title{Mathematical functions for \code{mmad_expr}}
\usage{
\method{Math}{mmad_expr}(x, ...)
}
\arguments{
\item{x}{An \code{mmad_expr}.}

\item{...}{Unused.}
}
\value{
An \code{mmad_expr}.
}
\description{
\code{log()}, \code{exp()}, and \code{sqrt()} (the latter as \code{(.)^(1/2)}).
}
