% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atoms.R
\name{mmad_atom_names}
\alias{mmad_atom_names}
\title{Names of currently registered atoms}
\usage{
mmad_atom_names()
}
\value{
A character vector.
}
\description{
Names of currently registered atoms
}
\examples{
mmad_atom_names()
}
\keyword{internal}
