% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr.R
\name{mmad_call}
\alias{mmad_call}
\title{Construct an atom-application expression node}
\usage{
mmad_call(op, args, params = list())
}
\arguments{
\item{op}{Name of a registered atom.}

\item{args}{List of \code{mmad_expr} children.}

\item{params}{Named list of additional numeric atom-parameters
(e.g. \code{list(c = 0.5)} for the exponent of a power atom).}
}
\value{
An object of class \code{mmad_expr} (subclass \code{mmad_call}).
}
\description{
Internal constructor; not normally called by users. Validates that the
atom name is registered (see \code{\link[=mmad_atom_names]{mmad_atom_names()}}) and that all child
expressions are themselves \code{mmad_expr}.
}
\keyword{internal}
