% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmr.R
\docType{package}
\name{mmr}
\alias{mmr}
\alias{mmr-package}
\title{mmr: A package for computating awareness measures for preclinical AD cohorts}
\description{
The mmr package provides a function that allows to compute two measures of
awareness: the Meta-Memory Ratio (MMR) and the Awareness of Cognitive
Decline Categorization (ACDC)
}
\section{Mmr functions}{

The mmr functions :

mmr_build: This function take as input a formated data.frame, as well as
    some mandatory entries (i.e. memory variables, complaint variables,
    vector of variables that have to be treated as normal or binomial).
    The output is the same data.frame, with the addition of the computed
    measure (i.e. MMR and ACDC).
    The mmr_build function is based on the mmr_reglin and mmr_regbin
    functions.

mmr_reglin: This function perform the prerequisite operations that allows the
    further computation of the MMR and ACDC. It stands for normally
    distributed variables.

mmr_regbin: This function is comparable to the mmr_reglin function but stands
    for variables that have to be treated as binomial.
}

