% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmr_acdilong.R
\name{mmr_acdilong}
\alias{mmr_acdilong}
\title{Compute the ACDI (Awareness for Cognitive Decline Index)}
\usage{
mmr_acdilong(self, ext, vis, df, lim = 1.5)
}
\arguments{
\item{self}{character (variable name for participant's complaint score)}

\item{ext}{character (variable name for informant's complaint score)}

\item{vis}{character (name of the 'visit' variable)}

\item{df}{data.frame}

\item{lim}{numeric (defined limits for group delineation. If unspecified,
lim = 1.5)}
}
\value{
data.frame. The return of the function corresponds to the previous
    data.frame (i.e. the one in the 'df' parameter) with the new computed values.
}
\description{
This index is based on the Participant-Informant Discrepancy. It computes
the difference between the participant's appraisal of his/her cognition
and the informant's.
}
