% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmr_build.R
\name{mmr_build}
\alias{mmr_build}
\title{Compute the Meta-Memory Ratio (MMR) and Awareness of Cognitive Decline Categorization (ACDC)}
\usage{
mmr_build(mem, comp, vect.linear, vect.binomial, df)
}
\arguments{
\item{mem}{character vector (memory variables)}

\item{comp}{character vector (complaint variables)}

\item{vect.linear}{character vector (normally distributed variables)}

\item{vect.binomial}{character vector (variables with binomial distribution)}

\item{df}{data.frame}
}
\value{
data.frame
}
\description{
Performs a regression on defined memory and complaint variables, and
standardized the residuals. After building composite memory and complaint
scores, the function compute both the MMR and the ACDC on this basis.
}
