% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmr_longit.R
\name{mmr_longit}
\alias{mmr_longit}
\title{Longitudinal Computation of the Meta-Memory Ratio (MMR) and Awareness of Cognitive Decline Categorization (ACDC)}
\usage{
mmr_longit(vis, mem, comp, vect.linear, vect.binomial, df)
}
\arguments{
\item{vis}{character (name of the 'visit' variable)}

\item{mem}{character vector (memory variables)}

\item{comp}{character vector (complaint variables)}

\item{vect.linear}{character vector (normally distributed variables)}

\item{vect.binomial}{character vector (variables with binomial distribution)}

\item{df}{data.frame}
}
\value{
data.frame. The return of the function corresponds to the previous
    data.frame (i.e. the one in the 'df' parameter) with the new computed values.
}
\description{
Performs a by visit regression on defined memory and complaint variables, and
standardized the residuals. After building composite memory and complaint
scores, the function compute both the MMR and the ACDC on this basis.
}
\details{
This function is based on one other mmr functions: mmr_build
}
\examples{


}
