% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmr_regbin.R
\name{mmr_regbin}
\alias{mmr_regbin}
\title{Cleaning of binomial variables}
\usage{
mmr_regbin(vect.binomial, df)
}
\arguments{
\item{vect.binomial}{character vector (variables with binomial distribution)}

\item{df}{data.frame()}
}
\value{
data.frame()
}
\description{
This function apply a regression (using glm()) on variables with
binomial distribution. Then the residuals are added in the original data.frame
}
