% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PETmm.R
\name{PETmm}
\alias{PETmm}
\title{Poisson model for transmission tomography}
\usage{
PETmm(theta, y, A, s, r)
}
\arguments{
\item{theta}{attenuation coefficient}

\item{y}{transmission measurement of detector}

\item{A}{the n*q system matrix}

\item{s}{the blank scan counts}

\item{r}{the mean number of background counts of detector}
}
\value{
requires a value
}
\description{
Poisson model for transmission tomography
}
\examples{
y=c(4,1); theta=c(3,-4);s=c(3.7, 3.8); r=c(0,0)
aa = c(0.029, 0.027, 0.007, 0.06)
A=matrix(aa,2)
PETmm(theta,y,A,s,r)
}
