% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.GaF.R
\name{plot.GaF}
\alias{plot.GaF}
\title{Plot the GaF object}
\usage{
\method{plot}{GaF}(
  x,
  XLAB = "Time",
  YLAB = "Cumulative hazard",
  TYPE = "s",
  LTY = 1,
  LWD = 1,
  COL = gray(0),
  digits = 4,
  ...
)
}
\arguments{
\item{x}{The GaF object, see \code{\link{GaFrailtyMM}}.}

\item{XLAB}{x label, default is "Time".}

\item{YLAB}{y label, default is "Cumulative hazard".}

\item{TYPE}{type value, default is "s".}

\item{LTY}{lty value for line, default is 1.}

\item{LWD}{line width, default is 1.}

\item{COL}{color parameter, default is gray(0).}

\item{digits}{The digits after the decimal point, default = 4.}

\item{...}{Additional arguments}
}
\value{
the dataframe of "Time" and accumulative hazard \eqn{\Lambda}.
}
\description{
Plot the GaF object
}
\examples{
library(survival)
result <- GaFrailtyMM(Surv(time, status) ~ age + sex + cluster(id), data=kidney)
\dontrun{ plot.GaF(result) }

}
