% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcp.R
\name{sign_of}
\alias{sign_of}
\title{Inferred sign of an \code{mmad_expr}}
\usage{
sign_of(expr)
}
\arguments{
\item{expr}{An \code{mmad_expr}.}
}
\value{
One of \code{"positive"}, \code{"nonneg"}, \code{"zero"}, \code{"nonpos"},
\code{"negative"}, \code{"unknown"}.
}
\description{
Inferred sign of an \code{mmad_expr}
}
\examples{
sign_of(exp(mmad_var(1)))                # "positive"
sign_of(mmad_var(1) ^ 2)                 # "nonneg"
}
