% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.LTN.R
\name{summary.LTN}
\alias{summary.LTN}
\title{Summary of parameter estimates of a LTN model}
\usage{
\method{summary}{LTN}(object, digits = 4, ...)
}
\arguments{
\item{object}{Output from a call to LTN.}

\item{digits}{The desired number of digits after the decimal point. Default of 4 digits is used.}

\item{...}{Additional arguments}
}
\value{
Summary for \code{LTNMM} objects.
}
\description{
This function returns the result of the \code{LTNMM} function
}
\examples{


y=c(8.7, 5.4, 8.9, 5.8, 6.2, 9.9, 7.5, 9.5, 6.5, 6.3); a=5
result <- LTNMM(y~1, a=5)

summary(result,digits=4)



}
\seealso{
\code{\link{LTNMM}}
}
\keyword{methods}
