% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDesc.R
\docType{data}
\name{vijc}
\alias{vijc}
\title{Voluntary and involuntary job changes data}
\format{
An object of class \code{data.frame} with 2124 rows and 2 columns.
}
\usage{
vijc
}
\value{
Voluntary and involuntary job changes data contains the following fields:
\item{y1}{Job changes after experiencing an unemployment spell(assumed to be involuntary)}
\item{y2}{Direct job to job changes(which are assumed to be voluntary) }
}
\description{
Jung and Winkelmann(1993) provided data on both the numbers of voluntary and involuntary job changes of males during ten period 1974–1984.
The samples contain 2124 males who started their working career before or in 1974 and did not retire before 1984.
}
\examples{
data = data(vijc)
}
\references{
Huang X.F., Tian G.L., Zhang, C. and Jiang, X.(2017). "Type I multivariate zero-inflated generalized Poisson distribution with applications." \emph{Statistics and its Interface} \strong{10}(2), 291-311.

Jung R.C. and Winkelmann R.(1993). "Two aspects of labor mobility: A bivariate Poisson regression approach." \emph{Empirical Economics} \strong{18}(3), 543–556.
}
\keyword{datasets}
