% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZIGP_CRate.R
\name{ZIGP_CRate}
\alias{ZIGP_CRate}
\title{Title}
\usage{
ZIGP_CRate(phi0, phi, la, th)
}
\arguments{
\item{phi0}{probability}

\item{phi}{probability}

\item{la}{scale parameter}

\item{th}{discrete parameter}
}
\value{

}
\description{
Title
}
\examples{
ZIGP_CRate(0.5,0.5,1,0.1)
}
