% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZIGP_Rate.R
\name{ZIGP_Rate}
\alias{ZIGP_Rate}
\title{Title}
\usage{
ZIGP_Rate(phi0, la, th)
}
\arguments{
\item{phi0}{probability}

\item{la}{scale parameter}

\item{th}{discrete parameter}
}
\value{
requires a value
}
\description{
Title
}
\examples{
ZIGP_Rate(0.5,1,0.1)
}
