% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZIGPsample.R
\name{ZIGP_sample}
\alias{ZIGP_sample}
\title{Generate sample from Type I Multivariate Zero-inflated Generalized Poisson Distribution}
\usage{
ZIGP_sample(phi0, la, th, n)
}
\arguments{
\item{phi0}{Probability of taking value 0 for Bernoulli distribution}

\item{la}{Lambda for Poison distribution}

\item{th}{Theta for Poison distribution}

\item{n}{Number of samples generated}
}
\value{
Vector of samples with dimension n
}
\description{
Generate sample from Type I Multivariate Zero-inflated Generalized Poisson Distribution
}
