% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZIGPFx.R
\name{ZIGPFx}
\alias{ZIGPFx}
\title{Title}
\usage{
ZIGPFx(phi0, la, th)
}
\arguments{
\item{phi0}{probability}

\item{la}{scale parameter}

\item{th}{discrete parameter}
}
\value{
requires a value
}
\description{
Title
}
\examples{
ZIGPFx(0.5,1,0.1)
}
