% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZIGPmm.R
\name{ZIGPmm}
\alias{ZIGPmm}
\title{MM algorithm based on the AD method for Multivariate compound zero-inflated generalized Poisson distribution}
\usage{
ZIGPmm(phi, la, th, y)
}
\arguments{
\item{phi}{probability}

\item{la}{scale parameter}

\item{th}{discrete parameter}

\item{y}{matrix data}
}
\value{
data
}
\description{
MM algorithm based on the AD method for Multivariate compound zero-inflated generalized Poisson distribution
}
\examples{
y = matrix(7,1,1); ZIGPmm(0.5,1,0.1,y)
}
