% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZIGPQx.R
\name{ZIGPQx}
\alias{ZIGPQx}
\title{Title}
\usage{
ZIGPQx(phi0, phi, la, th)
}
\arguments{
\item{phi0}{probability}

\item{phi}{probability}

\item{la}{scale parameter}

\item{th}{discrete parameter}
}
\value{

}
\description{
Title
}
\examples{
ZIGPQx(0.5,0.5,1,0.1)
}
