% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxmm.R
\name{CoxMM}
\alias{CoxMM}
\title{Title}
\usage{
CoxMM(
  formula,
  data,
  beta = NULL,
  Maxiter = 5000,
  convergence = 1e-06,
  bootstrap = TRUE,
  b = 100,
  ...
)
}
\arguments{
\item{data}{a data frame containing the variables of the model}

\item{Maxiter}{The maximum iterations}

\item{convergence}{convergence value}

\item{bootstrap}{bootstrap}

\item{b}{number of bootstrap iterations for the estimation of the standard errors}

\item{...}{}

\item{time}{for right censored data}

\item{status}{The status indicator, normally 0=alive, 1=dead.}

\item{be}{a vector of regression parameters}

\item{method}{Profile or Non Profile}
}
\value{
requires a value
}
\description{
Title
}
\examples{
data = matrix(rnorm(10*1,-1,1),10,1)
be = 0.5
u=runif(10); time=-log(u)/(la*exp(x\%*\%be))
cen=3.7; status=1*(t<=cen)
CoxMM(data, time, status, be)
}
