% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GaFrailtyMM.R
\name{GaFrailtyMM}
\alias{GaFrailtyMM}
\title{The MM algorithm based on AD technology for gamma frailty model}
\usage{
GaFrailtyMM(
  formula,
  data,
  beta = NULL,
  theta = NULL,
  lambda = NULL,
  Maxiter = 1000,
  convergence = 1e-06,
  ...
)
}
\arguments{
\item{formula}{A formula object, which contains on the left hand side an object of the type \code{Surv}
and on the right hand side a \code{+cluster(id)} statement.
e.g. \code{formula=Surv(time, status) ~ x + cluster(id)}}

\item{data}{A \code{data.frame} in which to interpret the variables named in the formula.}

\item{beta}{A vector of unknown regression parameters, default is \code{NULL}.
If is \code{NULL}, then make all \code{beta=1} during calculation.}

\item{theta}{The Variance of frailty factors subject to Gamma distribution, default is \code{NULL}.
If is \code{NULL}, then let \code{theta=1} during calculation.}

\item{lambda}{Baseline hazard rate, default set to \code{NULL}. If is \code{NULL},
then let each \code{lambda} equals to \code{1/N} during calculation, which \code{N} is the number of observed.}

\item{Maxiter}{The maximum number of iterations; default=1000.}

\item{convergence}{Specify the convergence criterion, the default is 1e-6.}

\item{...}{Additional arguments}
}
\value{
An object of class \code{GaFrailtyMM} that contains the following fields: Total amount of observations,
total number of failure events, total number of iterations, Convergence result, The Log Likelihood value,
variance of gamma distribution, The standard deviation of the estimated \eqn{\theta},
The likelihood-based 95\% confidence interval for the \eqn{\theta}, vector of regression parameters,
The standard deviation of the estimated \eqn{\beta}, The likelihood-based 95\% confidence interval for the \eqn{\beta},
The variable name.
}
\description{
The MM algorithm based on AD technology for gamma frailty model
}
\details{
The \code{GaFrailtyMM} function is used to fit Gamma frailty survival model using profile MM algorithms
based on AD technology. EM algorithms relies on the fact that, after profiling out the nonparametric component \eqn{\Lambda_0},
the resulting function is concave. When it does not hold, using Newton method to maximize the resulting function is
difficult especially when there exist a large number of covariates. The profile MM algorithms that can avoid the
concavity requirement and bypass Newton method and matrix inversion.
}
\examples{
GaFrailtyMM(Surv(time, status) ~ age + sex + cluster(id), data=kidney)

}
