% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IC2Control.R
\name{IC2Control}
\alias{IC2Control}
\title{control IC2Pro object}
\usage{
IC2Control(Maxiter = 2000, convergence = 1e-06, Idigits = 4, Pdigits = 4)
}
\arguments{
\item{Maxiter}{The maximum number of iterations; default=2000.}

\item{convergence}{Specify the convergence criterion, the default is 1e-6.}

\item{Idigits}{The digits for survival interval values.}

\item{Pdigits}{The digits for survival probability values.}
}
\value{
list of Maxiter, convergence, Idigits, Pdigits.
}
\description{
control IC2Pro object
}
\examples{
IC2Control()
}
