% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LTNmm.R
\name{LTNmm}
\alias{LTNmm}
\title{MM algorithm based on the AD method for Left-truncated normal distribution}
\usage{
LTNmm(
  y,
  a,
  mu = NULL,
  sigma = NULL,
  Maxiter = 2000,
  convergence = 1e-06,
  method = "ADMM",
  ...
)
}
\arguments{
\item{y}{data}

\item{a}{left border}

\item{mu}{mean}

\item{si2}{variance}
}
\value{
data
}
\description{
MM algorithm based on the AD method for Left-truncated normal distribution
}
\examples{
LTNmm(5, 7, 4, 3)
y=c(3, 9); LTNmm(5, 7, 4, y)
}
