% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LTNMM.R
\name{LTNMM}
\alias{LTNMM}
\title{MM algorithm based on the AD method for Left-truncated normal distribution}
\usage{
LTNMM(
  formula,
  a,
  mu = NULL,
  sigma = NULL,
  data = sys.frame(sys.parent()),
  Maxiter = 2000,
  convergence = 1e-06,
  ...
)
}
\arguments{
\item{formula}{A formula object which symbolically describes the model to calculated.}

\item{a}{A numeric scalar of the known left truncation value.}

\item{mu}{Mean of the normal distribution, default is NULL, if truncated only we use estimates from OLS.}

\item{sigma}{Variance of the normal distribution, default is NULL, if truncated only we use estimates from OLS.}

\item{data}{List that contains corresponding covariates. If none is provided then assumes objects are in user’s environment.}

\item{Maxiter}{The maximum number of iterations; default=2000.}

\item{convergence}{Specify the convergence criterion, the default is 1e-6.}

\item{...}{Additional arguments}
}
\value{
An object of class \code{LTNMM} that contains the following fields: total amount of observations,
the number of iterations, convergence rate, the Log Likelihood value, estimated results for the unknown parameters,
the standard deviation of estimate for the unknown parameters, the likelihood-based 95\% confidence interval for the unknown parameters,
Information Criterion: AIC value and BIC value.
}
\description{
The \code{LTNMM} function is used to calculate a left-truncated normal distribution model. A \eqn{ LTN(\mu, \sigma^2; a) } has the density function

\deqn{f(y; \mu, \sigma^2; a) = \frac{1}{c \sqrt{2 \pi \sigma^{2}}} \exp{( -\frac{(y-\mu)^{2}}{2 \sigma^{2}} )} \centerdot I(y \geq a) }

where \eqn{(\mu, \sigma^2)} are two unknown parameters, \eqn{ a } is a known constant, \eqn{ c = 1- \Phi(\frac{a-u}{ \sigma}) }, and  \eqn{ \Phi(\centerdot) } is the cdf of the standard normal distribution.
}
\details{
The \code{LTNMM} function is used to fit Left-truncated normal distribution model using MM algorithms
based on AD technology. \code{formula} can be used to provide data that needs to be fitted, such as \code{formula=y~1}.
\code{data} is provided by user’s environment by default. The initial value of mean and variance of the normal distribution estimated form OLS.
}
\examples{
y=c(8.7, 5.4, 8.9, 5.8, 6.2, 9.9, 7.5, 9.5, 6.5, 6.3); a=5
LTNMM(y~1, a=5)

}
\references{
Tian G.L., Huang X.F., and Xu, J.(2019). "An assembly and decomposition approach for constructing separable minorizing functions in a class of MM algorithms." \emph{Statistica Sinica} \strong{29}(2), 961-982.
}
