% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.IC2.R
\name{plot.IC2}
\alias{plot.IC2}
\title{Plot the IC2 object}
\usage{
\method{plot}{IC2}(
  x,
  XLAB = "Time",
  YLAB = "Survival",
  LEGEND = NULL,
  MAIN = "Survival Function",
  LTY = 1:9,
  LWD = 1,
  XLEG = 0,
  YLEG = 0.15,
  COL = gray(0),
  ...
)
}
\arguments{
\item{x}{The IC2 object, see \code{\link{IC2MM}}.}

\item{XLAB}{x label, default is "Time".}

\item{YLAB}{y label, default is "Survival".}

\item{LEGEND}{legend, default=NULL.}

\item{MAIN}{figure title, default is "Survival Function"}

\item{LTY}{lty value for line, default is 1:9.}

\item{LWD}{line width, default is 1.}

\item{XLEG}{positional parameters of the legend, default=0.}

\item{YLEG}{positional parameters of the legend, default=0.15 .}

\item{COL}{the color of the drawing, default=gray(0)}

\item{...}{Additional arguments}
}
\value{
the Survival Function plot
}
\description{
Plot the IC2 object
}
\examples{

library(survival)
result = IC2MM(Surv(left, right, type = "interval2") ~ treatment, bcos)
\dontrun{
summary.IC2(result)
plot.IC2(result, COL=c("red", "blue"))
}

}
