% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.Cox.R
\name{summary.Cox}
\alias{summary.Cox}
\title{summary of parameter estimates of a Cox model}
\usage{
\method{summary}{Cox}(object, digits = 4, ...)
}
\arguments{
\item{object}{Output from a call to Cox.}

\item{digits}{The desired number of digits after the decimal point. Default of 4 digits is used.}

\item{...}{Additional arguments}
}
\value{
Prints the fitting result of the \code{CoxMM} function.
}
\description{
This function returns the fitting result of the \code{CoxMM} function
}
\examples{

library(survival)
result <- CoxMM(Surv(time, status) ~ age + sex, lung)
\dontrun{
summary.Cox(result,digits=4)
}


}
\seealso{
\code{\link{CoxMM}}
}
\keyword{methods}
