% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.IC2.R
\name{summary.IC2}
\alias{summary.IC2}
\title{Summary of parameter estimates of a IC2 model}
\usage{
\method{summary}{IC2}(object, ...)
}
\arguments{
\item{object}{Output from a call to IC2.}

\item{...}{Additional arguments}
}
\value{
Summary for \code{IC2MM} objects.
}
\description{
This function returns the result of the \code{IC2MM} function
}
\examples{


library(survival)
result <- IC2MM(Surv(left, right, type = 'interval2') ~ treatment, bcos)

summary(result)



}
\seealso{
\code{\link{IC2MM}}
}
\keyword{methods}
