% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZIGPmm.R
\name{ZIGPmm}
\alias{ZIGPmm}
\title{Title}
\usage{
ZIGPmm(
  y,
  phi0,
  la,
  th,
  Maxiter = 500,
  convergence = 1e-06,
  method = "ADMM",
  ...
)
}
\arguments{
\item{y}{matrix data}

\item{phi0}{probability}

\item{la}{scale parameter}

\item{th}{discrete parameter}

\item{Maxiter}{The maximum number of iterations}

\item{convergence}{Convergence value}

\item{method}{optimized method}

\item{...}{}
}
\value{
requires a value
}
\description{
Title
}
\examples{
y = matrix(7,1,1);
ZIGPmm(y,0.5,1,0.1)
}
