% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZTB1mm.R
\name{ZTB1mm}
\alias{ZTB1mm}
\title{MM algorithm based on the AD method for Zero-truncated binomial distribution(1)}
\usage{
ZTB1mm(m, th, y)
}
\arguments{
\item{m}{number of experiments}

\item{th}{probability of success}

\item{y}{}
}
\value{

}
\description{
MM algorithm based on the AD method for Zero-truncated binomial distribution(1)
}
\examples{
y = c(1, 2)
ZTB1mm(5, 0.2, y)
}
